move dialog base
